refactor add licence
