+++ .gitignore
@@ -17,3 +17,6 @@ doc/
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+.idea/
+*.iml
